+++ src/test/java/kitchenpos/acceptance/ProductAcceptanceTest.java
@@ -6,7 +6,7 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.acceptance.helper.KitchenPosBehaviors;
+import kitchenpos.utils.KitchenPosBehaviors;
 import org.junit.jupiter.api.DisplayName;
 import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.Test;
