+++ src/test/java/kitchenpos/acceptance/TableGroupAcceptanceTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.acceptance.helper.KitchenPosBehaviors;
+import kitchenpos.utils.KitchenPosBehaviors;
 import kitchenpos.fixture.OrderTableFixtureFactory;
 import kitchenpos.fixture.TableGroupFixtureFactory;
 import kitchenpos.table.dto.OrderTableResponse;
