--- src/test/java/kitchenpos/domain/table/OrderTableTest.java
@@ -1,5 +1,0 @@
-package kitchenpos.domain.table;
-
-class OrderTableTest {
-
-}
