--- src/test/java/kitchenpos/domain/table/TableGroupTest.java
@@ -1,4 +1,0 @@
-package kitchenpos.domain.table;
-
-class TableGroupTest {
-}
