+++ src/test/java/kitchenpos/table/domain/OrderTableTest.java
@@ -1,0 +1,4 @@
+package kitchenpos.table.domain;
+class OrderTableTest {
+
+}
