+++ src/test/java/kitchenpos/table/domain/TableGroupTest.java
@@ -1,0 +1,4 @@
+package kitchenpos.table.domain;
+
+class TableGroupTest {
+}
