+++ src/test/java/kitchenpos/utils/KitchenPosBehaviors.java
@@ -1,4 +1,4 @@
-package kitchenpos.acceptance.helper;
+package kitchenpos.utils;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
